% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajout_libelles_varventil_insee.R
\name{ajout_libelles_varventil_insee}
\alias{ajout_libelles_varventil_insee}
\title{Ajouter le libellé d'une modalité de variable de ventilation}
\usage{
ajout_libelles_varventil_insee(TABLE, VAR, MILLESIME_RP)
}
\arguments{
\item{TABLE}{Table en entrée avec un champ contenant les modalités d'une variable de ventilation en format court (liste : cf. details ci-dessous).}

\item{VAR}{Nom du champ contenant les modalités d'une variable de ventilation en format court (liste : cf. details ci-dessous).}

\item{MILLESIME_RP}{Millésime du recensement RP de l'Insee de la variable de ventilation en entrée (liste : cf. details ci-dessous).}
}
\value{
Renvoie une table avec un champ supplémentaire contenant le libellé de la modalité.
}
\description{
Ajouter le libellé d'une modalité de variable de ventilation (issue d'un fichier du RP Insee) à partir du champ contenant le code en format court.
}
\examples{
# Ajout du champ 'LIB_CS1' contentant le libellé en clair du code CS1 (aka "groupe socio-professionnel")
indics_mig_DEP_CS1 <-
ajout_libelles_varventil_insee(TABLE = indics_mig_DEP_CS1,
                               VAR ="CS1",
                               MILLESIME_RP = 2014)



}
