% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajout_nivgeo_supracomm.R
\name{ajout_nivgeo_supracomm}
\alias{ajout_nivgeo_supracomm}
\title{Ajouter un champ contenant un niveau géographique supra-communal}
\usage{
ajout_nivgeo_supracomm(TABLE, CODE_COMMUNE, SUFFIXE = "", COG_IN,
  COG_NIVGEO, NIVGEO)
}
\arguments{
\item{TABLE}{Table en entrée disposant a minima d'un champ de codes commune Insee.}

\item{CODE_COMMUNE}{Nom du champ de la table en entrée contenant le code commune Insee. Il peut contenir des codes arrondissements pour Paris, Lyon et Marseille (75101, 75102, etc...) ou non ('75056')}

\item{SUFFIXE}{Terme à ajouter en suffixe au nom du champ de niveau supra-communal ajouté à la table (facultatif).}

\item{COG_IN}{Millésime du Code Officiel Géographique du code commune en entrée (liste : cf. details ci-dessous). Le millésime du COG est indiqué dans les métadonnées des fichiers Insee, dans le fichier du RP d'année N il est généralement égal à N-2.}

\item{COG_NIVGEO}{Millésime du Code Officiel Géographique du code supra-communal en sortie (liste : cf. details ci-dessous).}

\item{NIVGEO}{Nom du niveau géographique supra-communal à ajouter (en format court, liste : cf. details ci-dessous). Peut être un vecteur contenant les noms de plusieurs niveaux.}
}
\value{
Renvoie une table avec un champ supplémentaire contenant le niveau géographique supra-communal.
}
\description{
Dans une table contenant un code commune Insee (avec ou sans arrondissements pour Paris, Lyon et Marseille), accoler à cette table un nouveau champ contenant le niveau supra-communal correspondant à cette commune.
}
\examples{
# Ajout du champ 'DEP' indiquant le niveau géographique "département" à une table ayant une information communale
COM_2012 <-
ajout_nivgeo_supracomm(TABLE = COM_2012 ,
CODE_COMMUNE = "CODGEO",
NIVGEO = "DEP",
COG_IN = 2010,
COG_NIVGEO = 2018)

# Ajout des niveaux géographiques "département", "région" et "type de densité" depuis les champs "commune de résidence actuelle" et "commune de résidence antérieure"
FD_MIGCOM_2014  <-
ajout_nivgeo_supracomm(TABLE = FD_MIGCOM_2014 ,
                       CODE_COMMUNE = "DCRAN",
                       SUFFIXE = "ANTE",
                       NIVGEO= c("DEP",'EPCI','TYPEDENS'),
                       COG_IN = 2016, COG_NIVGEO = 2018) \%>\%
ajout_nivgeo_supracomm(TABLE = . ,
                       CODE_COMMUNE = "COMMUNE",
                       SUFFIXE = "ACTU",
                       NIVGEO= c("DEP",'EPCI','TYPEDENS'),
                       COG_IN = 2016, COG_NIVGEO = 2018)

}
\references{
\itemize{
\item{\href{https://antuki.github.io/COGugaison/articles/COGugaison.html}{Package 'COGugaison' pour gérer les changements de COG}}
\item{\href{https://www.insee.fr/fr/information/2666684#titre-bloc-11}{Historique des géographies communales (Insee)}}
\item{\href{https://www.insee.fr/fr/information/2028028}{Tables d'appartenance des communes aux différents niveaux géographiques (Insee)}}}
}
