% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionalPrices.r
\name{both_prices_output}
\alias{both_prices_output}
\title{Produce Table 6 in Main Text}
\usage{
both_prices_output()
}
\description{
Collects R2 from all regressions in \code{\link{reg_vs_state_y}} and \code{\link{reg_vs_state_p}}
}
