% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_flux_migres.R
\name{calcul_flux_migres}
\alias{calcul_flux_migres}
\title{Calculer le volume de migrations résidentielles entre zones}
\usage{
calcul_flux_migres(TABLE, NIVGEO_ANTE, NIVGEO_ACTU, VAR_NB,
  VAR_VENTIL = NULL, MIG_NET_INTERNE = TRUE)
}
\arguments{
\item{TABLE}{Table en entrée contenant un champ de résidence actuelle et un champ de résidence antérieure sur le même niveau géographique (communes, départements, etc...).}

\item{NIVGEO_ANTE}{Nom du champ de la table en entrée contenant le niveau géographique de résidence antérieure.}

\item{NIVGEO_ACTU}{Nom du champ de la table en entrée contenant le niveau géographique de résidence actuelle.}

\item{VAR_NB}{Nom du champ de la table en entrée contenant l'indicateur de pondération du tuple.}

\item{VAR_VENTIL}{Nom du champ de la variable de ventilation de la population, de type tranche d'âge, groupe socio-professionnel, etc... (facultatif).}

\item{MIG_NET_INTERNE}{Vaut TRUE si le calcul doit se faire uniquement sur les zones présentes à la fois en résidence actuelle et en résidence antérieure.}
}
\value{
Renvoie une table contenant un champ de résidence actuelle, un champ de résidence antérieure et un champ du nombre de migrations résidentielles entre chaque coupe de zones.
}
\description{
A partir d'une table en entrée contenant un champ de résidence actuelle et un champ de résidence antérieure sur le même niveau géographique.
}
\examples{
# Calcul de la table de nombre de migrations entre résidence actuelle et résidence antérieure au niveau départemental
flux_migres_DEP <-
calcul_flux_migres(TABLE =FD_MIGCOM_2014 ,
                 VAR_NB = "IPONDI",
                 MIG_NET_INTERNE=TRUE,
                 NIVGEO_ANTE ="DEP_ANTE",
                 NIVGEO_ACTU ="DEP_ACTU")
# Calcul de la table de nombre de migrations entre résidence actuelle et résidence antérieure au niveau régional, et ventilés par groupe socio-professione
flux_migres_REG_CS1 <-
calcul_flux_migres(TABLE =FD_MIGCOM_2014 ,
                 VAR_NB = "IPONDI",
                 VAR_VENTIL = "CS1",
                 MIG_NET_INTERNE=TRUE,
                 NIVGEO_ANTE ="REG_ANTE",
                 NIVGEO_ACTU ="REG_ACTU")

}
