% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_indics_migres.R
\name{calcul_indics_migres}
\alias{calcul_indics_migres}
\title{Calculer les indicateurs relatifs aux migrations résidentielles pour chaque zone.}
\usage{
calcul_indics_migres(TABLE, NIVGEO_ACTU, NIVGEO_ANTE, NIVGEO, VAR_NB,
  VAR_VENTIL = NULL)
}
\arguments{
\item{TABLE}{Table en entrée contenant a minima un champ de résidence actuelle et un champ de résidence antérieure sur le même niveau géographique, ainsi qu'un champ comptablisant le nombre d'individus correspondant.}

\item{NIVGEO_ACTU}{Nom du champ de la table en entrée contenant le niveau géographique de résidence actuelle.}

\item{NIVGEO_ANTE}{Nom du champ de la table en entrée contenant le niveau géographique de résidence antérieure.}

\item{NIVGEO}{Nom du champ de la table en sortie contenant le niveau géographique.}

\item{VAR_NB}{Nom du champ de la table en entrée contenant le nombre d'individus correspondant à la migration résidence antérieure / résidence actuelle.}

\item{VAR_VENTIL}{Nom du champ de la variable de ventilation de la population, de type tranche d'âge, groupe socio-professionnel, etc... (facultatif).}
}
\value{
Renvoie une table contenant un champ contenant l'identifiant des territoires ainsi que les indicateurs relatifs aux migrations résidentielles.
}
\description{
A partir d'une table en entrée contenant un champ de résidence actuelle et un champ de résidence antérieure sur le même niveau géographique, une table en sortie est créée contenant les indicateurs relatifs aux migrations résidentielles (liste des indicateurs : cf. details ci-dessous).
}
\details{
Les indicateurs générés dans la table de sortie sont les suivants (cf. vignette pour les définitions exactes et les formules de calcul) : \cr
\itemize{
\item{Population présente ('nb_ind_PRES')}
\item{Population autochtone ('nb_ind_AUTO')}
\item{Population stable ('nb_ind_ISO')}
\item{Population entrante ('nb_ind_ENTR')}
\item{Population sortante ('nb_ind_SORT')}
\item{Solde migratoire ('SM')}
\item{Taux de migration nette interne ('TM')}
\item{Taux de rotation nette interne ('TR')}
\item{Part d'entrants ('PE')}
\item{Part de sortants ('PS')}}
}
\examples{
# Calcul du nombre de migrations entre résidence actuelle et résidence antérieure au niveau départemental
indics_migres_REG <-
 calcul_indics_migres(TABLE =flux_migres_DEP,
                      NIVGEO_ANTE ="DEP_ANTE",
                      NIVGEO_ACTU ="DEP_ACTU",
                      NIVGEO ="DEP",
                      VAR_NB = "nb_ind")

# Calcul du nombre de flux entre résidence actuelle et résidence antérieure au niveau régional, et ventilés par groupe socio-professionel
indics_migres_REG_CS1 <-
 calcul_indics_migres(TABLE =flux_migres_REG_CS1,
                      NIVGEO_ANTE ="DEP_ANTE",
                      NIVGEO_ACTU ="DEP_ACTU",
                      NIVGEO ="DEP",
                      VAR_NB = "nb_ind")


}
