% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_indics_renouv.R
\name{calcul_indics_renouv}
\alias{calcul_indics_renouv}
\title{Calculer les indicateurs de catégorisation et de renouvellement de la population par les migrations}
\usage{
calcul_indics_renouv(TABLE, NIVGEO, NB_ENTR, NB_SORT, NB_AUTO, NB_PRES,
  VAR_VENTIL)
}
\arguments{
\item{TABLE}{Table en entrée contenant les indicateurs relatifs aux migrations résidentielles par territoire (population entrante, sortante, autochtone et présente).}

\item{NIVGEO}{Nom du champ de la table en entrée contenant le niveau géographique.}

\item{NB_ENTR}{Nom du champ de la table en entrée contenant le nombre d'individus étant entré dans le territoire.}

\item{NB_SORT}{Nom du champ de la table en entrée contenant le nombre d'individus étant sorti du territoire.}

\item{NB_AUTO}{Nom du champ de la table en entrée contenant le nombre d'individus étant présente dans le territoire en N-x.}

\item{NB_PRES}{Nom du champ de la table en entrée contenant le nombre d'individus étant présente dans le territoire en N.}

\item{VAR_VENTIL}{Nom du champ de la variable de ventilation de la population, de type tranche d'âge, groupe socio-professionnel, etc... (facultatif).}
}
\value{
Renvoie une table contenant l'identifiant des territoires ainsi que les indicateurs par type de migrations (immigration, émigration, migrations) par modalité de la variable de ventilation (groupe socio-professionnel par exemple)
}
\description{
A partir d'une table en entrée contenant les indicateurs relatifs aux migrations résidentielles par territoire (population entrante, sortante, autochtone et présente).
}
\examples{

# Calcul les indicateurs de catégorisation et de renouvellement de la population par les migrations selon le groupe socio-professionnel au niveau régional
indics_mig_EPCI_CS1_RENOUV <-
 calcul_indics_renouv(TABLE = indics_mig_EPCI_CS1,
                      NIVGEO = "EPCI",
                      NB_ENTR = "nb_ind_ENTR",
                      NB_SORT = "nb_ind_SORT",
                      NB_AUTO = "nb_ind_AUTO",
                      NB_PRES = "nb_ind_PRES",
                      VAR_VENTIL ="CS1")




}
