% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chargement_fd_migcom.R
\name{chargement_fd_migcom}
\alias{chargement_fd_migcom}
\title{Charger un fichier détail MIGCOM de l'Insee}
\usage{
chargement_fd_migcom(telechargement = FALSE, anneeRP, dossier_dest_TL,
  chemin_FD)
}
\arguments{
\item{telechargement}{Vaut TRUE si le fichier doit être téléchargé depuis le site internet de l'Insee, FALSE si le fichier txt est disponible en local.}

\item{anneeRP}{Millésime du recensement de la population (RP) de l'Insee. Peut prendre les valeurs "2008", "2013" ou "2014". Non renseigné si telechargement vaut TRUE.}

\item{dossier_dest_TL}{Dossier de destination où va être stockée l'archive zip et les fichiers txt dézippés. S'il n'existe pas, il sera créé. Non renseigné si telechargement vaut FALSE.}

\item{chemin_FD}{Lien vers le fichier txt (FD_MIGCOM) en local. Non renseigné si telechargement vaut TRUE.}
}
\value{
Retourne une table de détail des migrations résidentielles avec pour chaque individu sa commune de résidence actuelle et commune de résidence antérieure.
}
\description{
Télécharger un fichier détail MIGCOM (Migrations résidentielles entre commune de résidence et commune de résidence antérieure) depuis le site de l'Insee puis le monter en mémoire, ou uniquement le charger si le fichier txt est disponible en local.
}
\examples{
# Téléchargement du fichier MIGCOM du dernier RP en date depuis le site internet de l'Insee dans le dossier "./data"
FD_MIGCOM_2014 <- chargement_fd_migcom(anneeRP = "2014",  telechargement = TRUE, dossier_dest_TL = "./data")

# Chargement du fichier txt MIGCOM disponible en local
FD_MIGCOM_2014 <- chargement_fd_migcom(chemin_FD = "./data/rp2014_migcom_txt/FD_MIGCOM_2014.txt")
}
