% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPS.r
\name{Clean.CPS}
\alias{Clean.CPS}
\title{Clean CPS data}
\usage{
Clean.CPS(dta = "~/datasets/CPS/outdata/selected.dta")
}
\description{
Clean CPS data
}
\details{
cleans CPS downloaded from \url{http://www.nber.org/data/current-population-survey-data.html}
I use the 2013 March supplement, documentation here \url{http://www.nber.org/cps/cpsmar13.pdf}
}
