% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotjulia.R
\name{combine_BEA_fhfa}
\alias{combine_BEA_fhfa}
\title{Get Macro Price Series}
\usage{
combine_BEA_fhfa()
}
\description{
Get Macro Price Series
}
\details{
This function writes out a dataset which contains
the series q and p used in the paper for each Division.
This function uses output from \code{\link{get_BEA_persincome}} and
\code{\link{getFHFA_realPrices}}.
}
