% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotjulia.R
\name{correlograms}
\alias{correlograms}
\title{Regional Price Correlograms}
\usage{
correlograms(path = "~/Dropbox/research/mobility/output/data/FHFA")
}
\arguments{
\item{path}{file path to output figure}
}
\value{
Produces figure 2 in main text.
}
\description{
Regional Price Correlograms
}
\details{
Uses regional time series on q and p to illustrate
the structure of the joint process (q,p) across regions.
}
