% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPS.r
\name{CPS.distance}
\alias{CPS.distance}
\title{Main Reason to Move: CPS data}
\usage{
CPS.distance(path = "~/Dropbox/mobility/output/data/cps")
}
\arguments{
\item{path}{to output file}
}
\value{
produces table 2 in the main text.
}
\description{
Main Reason to Move: CPS data
}
\details{
Uses CPS data cleaned in \code{\link{Clean.CPS}} to produce
a frequency table showing the main reasons to move.
}
