% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SippPrepare.r
\name{download.FHFA}
\alias{download.FHFA}
\title{get FHFA state level HPI}
\usage{

  download.FHFA(from = "http://www.fhfa.gov/webfiles/25831/3q13hpists_expandeddata.txt",
  to = "~/git/migration/mig-pkg/data")
}
\arguments{
\item{from}{url}

\item{to}{location to save data}
}
\value{
TRUE
}
\description{
download FHFA state-level house price
index 1990-2013, quarterly data.
}
\details{
source of data is
\url{http://www.fhfa.gov/Default.aspx?Page=87}
}
