% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotjulia.R
\name{Export.IncomeProcess}
\alias{Export.IncomeProcess}
\title{Estimate and Export Individual Income Process}
\usage{
Export.IncomeProcess(dat, writedisk, nocollege = FALSE,
  path = "~/Dropbox/research/mobility/output/model/fit")
}
\arguments{
\item{dat}{A data.table of SIPP micro data}

\item{writedisk}{boolean whether to save to disk}

\item{nocollege}{boolean whether subsetting to no college degree or not.}

\item{path}{string for graph output location}
}
\value{
Implements equation (21) in main text, writes table C.1 in online appendix to disk and produces figure C.1 titled \emph{Labor Income profiles for different q levels} also in online appendix.
}
\description{
Estimate and Export Individual Income Process
}
\details{
Uses SIPP micro data to estimate an income process, which is used in the structural model to predict and simulate individual income. This is the implementation of equation (21) in the main text, further illustrated in online appendix C.1. 
This function is called from \code{\link{Export.Julia}}, so please refer for input arguments to that function.
The procedure subsets sipp income data to leave out year 2007, which I found to be full of inconsistencies.
}
