% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotjulia.R
\name{Export.Julia}
\alias{Export.Julia}
\title{Export Data to Julia Package}
\usage{
Export.Julia(writedisk = TRUE, noCollege = TRUE)
}
\arguments{
\item{writedisk}{boolean TRUE whether to save to disk}

\item{noCollege}{boolean TRUE whether to subset data to individuals without college degree.}
}
\description{
Export Data to Julia Package
}
\details{
Takes all output data from this R package (mainly: moments)
and stores them on disk such that they can be used to run the structural
julia model in \code{migration/mig/}.
}
