% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotjulia.R
\name{Export.VAR}
\alias{Export.VAR}
\title{Export Aggregate/Regional VAR Processes}
\usage{
Export.VAR(plotpath = "~/Dropbox/research/mobility/output/data/sipp",
  writedisk)
}
\arguments{
\item{plotpath}{file path to output location}

\item{writedisk}{boolean whether to save to disk}
}
\value{
produces \enumerate{
\item Table 5 in main text \emph{Estimates for Aggregate VAR process}
\item Table B.4 in online appendix \emph{Aggregate to Regional price mappings}
\item Figure 3 in main text \emph{VAR fit to regional price data (p)}
\item Figure B.3 \emph{VAR fit to regional productivity data (q)}
\item Figures B.1 and B.2 showing the raw data series, titled \emph{Regional (p) and National (P) house price index} and \emph{Regional (q) and National (Q) Labor Productivity index}
\item returns all parameters of the estimated models
}
}
\description{
Export Aggregate/Regional VAR Processes
}
