% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SippPrepare.r
\name{Extract.wrap}
\alias{Extract.wrap}
\title{Extractor wrapper}
\usage{
Extract.wrap(verbose = TRUE, which = paste0(c(1996, 2001, 2004, 2008)),
  dblocation = "~/datasets/SIPP/R",
  dropbox = "C:/Users/florian_o/Dropbox/mobility/data/SIPP")
}
\arguments{
\item{verbose}{}

\item{which}{names of datasets to extract. Names are: "1996", "2001", "2004", "2008", "Mig_2008"}

\item{dblocation}{path to location of SIPP database, obtained as illustrated in \url{https://github.com/floswald/asdfree/blob/master/SIPP/down1996.R}.}

\item{dropbox}{path to folder where to save this}
}
\value{
NULL saves subset data.tables into dropbox
}
\description{
Selects variables from SIPP database
and does some initial subsetting. This
is an interface to \code{\link{ExtractorSippDB}}
}
