% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SippPrepare.r
\name{get.istate}
\alias{get.istate}
\title{auxiliary function to get movers
origin and destination state in a data.table (AT THE BEGINNING OF CURRENT PERIOD, SAY)}
\usage{
get.istate(states, imove)
}
\description{
auxiliary function to get movers
origin and destination state in a data.table (AT THE BEGINNING OF CURRENT PERIOD, SAY)
}
\examples{
ttab = data.table(pid = rep(c(1,2),each=5),state=c(3,3,4,4,4,6,7,7,8,9),istate=c(FALSE,FALSE,TRUE,FALSE,FALSE,FALSE,TRUE,FALSE,TRUE,TRUE))
ttab[,c("from","to") := get.istate(states=state,imove=istate),with=FALSE]
}
