% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotjulia.R
\name{get_BEA_persincome}
\alias{get_BEA_persincome}
\title{Get Personal Income Data from BEA}
\usage{
get_BEA_persincome()
}
\description{
Get Personal Income Data from BEA
}
\details{
this function depends on R package EconData available at
\url{https://github.com/floswald/EconData}
}
