% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotjulia.R
\name{getFHFA_realPrices}
\alias{getFHFA_realPrices}
\title{Extend SIPP/FHFA data back to 1967}
\usage{
getFHFA_realPrices()
}
\description{
Extend SIPP/FHFA data back to 1967
}
\details{
This takes the mean house value by division in base year 2012
and uses several price indices to project this value backwards in time.
\enumerate{
\item FHFA Division Index goes back until 1975
\item Use CPI to cover 1967-1975
}
}
