% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SippPrepare.r
\name{merge.idx}
\alias{merge.idx}
\title{Merge SIPP cores and topical modules auxiliary function}
\usage{
\method{merge}{idx}(core, topic, breaks = c(3, 6, 9, 12),
  topic.names = NULL)
}
\arguments{
\item{core}{list of core datasets}

\item{topic}{list of topical datasets}

\item{breaks}{numeric vector of waves where a TM was asked.}

\item{topic.names}{NULL by default assumes names of 
\code{topic} are like "TM_2". if not, supply names here.}
}
\description{
merges the most recent wealth module onto the corresponding
core data.  i.e. if breaks=c(3,6,9), the TM was asked
in waves 3,6 and 9. therefore merge TM_3 onto cores
1-3, merge TM_6 onto 4-6, etc
}
\examples{
co <- lapply(1:12, function(x) data.table(ssuid=1:4,covar=rnorm(4),key="ssuid"))
br <- c(2,5,9,12)
tm <- lapply(1:5, function(x) data.table(ssuid=1:4,tmvar=10*c(1,br)[x] + sample(1:4,size=4),key="ssuid"))
names(tm) <- paste0("TM_",c(1,br))	# don't merge first TM
merge.idx(core=co,topic=tm,breaks=br)
}
