% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migR-package.r
\name{migR}
\alias{migR}
\title{migR}
\description{
R companion package for \emph{The Effect of Homeownership on the Option Value of Regional Migration}
forthcoming in Quantitative Econmics, \url{www.qeconomics.org}.
}
\author{
Florian Oswald <Florian.oswald@gmail.com>
}
