% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{list of ggplots}

\item{cols}{Number of columns in layout}

\item{layout}{matrix specifying the layout. If present, 'cols' is ignored.
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.}
}
\description{
Multiple plot function
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
}
\author{
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
