% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotjulia.R
\name{plot_moment_fit}
\alias{plot_moment_fit}
\title{Plot Data vs Model Moments}
\usage{
plot_moment_fit(path = "~/Dropbox/research/mobility/output/model/fit")
}
\arguments{
\item{path}{location where julia objective function saves moments in file \code{moms.json}}
}
\value{
produces \enumerate{
\item Figure 4 in main text \emph{Graphical device to show model fit}
\item Figure D.1 in appendix \emph{Auxiliary Models} and \emph{Wealth}
}
}
\description{
Plot Data vs Model Moments
}
\details{
takes structural model-generated moments and compares them to data moments in a scatter plot.
}
