% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{PlotSippMigrationRates}
\alias{PlotSippMigrationRates}
\title{Plot Sipp Migration Rates by Age and Ownership Status}
\usage{
PlotSippMigrationRates(nocollege = FALSE)
}
\arguments{
\item{nocollege}{Boolean whether to subset to no college population.}
}
\value{
Figure 1 in the main text.
}
\description{
Plot Sipp Migration Rates by Age and Ownership Status
}
\details{
Generates a scatter plot of mobility vs age and a plot showing ownership by age.
}
