% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{PlotSippTransitionMatrix}
\alias{PlotSippTransitionMatrix}
\title{Plot Sipp Transition matrix}
\usage{
PlotSippTransitionMatrix(ttable,
  path = "~/Dropbox/mobility/output/data/sipp")
}
\description{
Plot Sipp Transition matrix
}
\examples{
\dontrun{
load('~/Dropbox/mobility/SIPP/Sipp_aggby_age.RData')
tt <- merged[from!=to,table(from,to)]
PlotSippTransitionMatrix(tt)
}
}
