% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionalPrices.r
\name{reg_vs_state_p}
\alias{reg_vs_state_p}
\title{Compare State vs Division Level p Indices}
\usage{
reg_vs_state_p()
}
\description{
Compare State vs Division Level p Indices
}
\details{
produces table B.6 in online appendix
}
