% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionalPrices.r
\name{reg_vs_state_y}
\alias{reg_vs_state_y}
\title{Compare State vs Division Level q Indices}
\usage{
reg_vs_state_y()
}
\description{
Compare State vs Division Level q Indices
}
\details{
produces table B.7 in online appendix
}
