% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SippAnalyse.r
\name{Sipp.moments}
\alias{Sipp.moments}
\title{Calculate Moments from SIPP data}
\usage{
Sipp.moments(d, svy, ages = c(20, 50))
}
\arguments{
\item{d}{a SIPP data.table}

\item{svy}{a SIPP dataset in survey format. Can be obtained by calling \code{\link{SippSvyDesign}}}
}
\value{
a datatable with moment names, values and standard deviations
}
\description{
Calculate Moments from SIPP data
}
\details{
Computes moments form SIPP data to be used in
structural model SMM estimation routine
}
