% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SippAnalyse.r
\name{Sipp.own_in_j_rent_in_k}
\alias{Sipp.own_in_j_rent_in_k}
\title{Fraction of non-resident landlords}
\usage{
Sipp.own_in_j_rent_in_k(path = "~/Dropbox/mobility/output/data/sipp")
}
\arguments{
\item{path}{to save graphs}
}
\description{
How many owners live in region k while renting out
a flat in region d?
}
