% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SippAnalyse.r
\name{Sipp.SumStats}
\alias{Sipp.SumStats}
\title{SIPP Summary Statistics}
\usage{
Sipp.SumStats(path = "~/Dropbox/research/mobility/output/data/sipp",
  nocollege = FALSE)
}
\arguments{
\item{path}{to save output}

\item{nocollege}{boolean whether to subset stats to no college degree.}
}
\value{
writes table 3 in main text to disk (\emph{Annual moving rate in percent of the population}).
}
\description{
SIPP Summary Statistics
}
\details{
Produces summary statistics from SIPP micro dataset
}
