% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SippAnalyse.r
\name{Sipp.wage_residual_copulas}
\alias{Sipp.wage_residual_copulas}
\title{Estimation of Movers' z Copula}
\usage{

  Sipp.wage_residual_copulas(path = "~/Dropbox/research/mobility/output/data/sipp")
}
\arguments{
\item{path}{to save graphs}
}
\value{
\enumerate{
\item Table C.2 in online appendix reporting estimated Copula parameters.
\item Figure C.2 illustrating the marginal distributions of z in periods before and after move
\item Figure C.3 is produced in the julia package, not here.
}
}
\description{
Estimation of Movers' z Copula
}
\details{
Implements the estimation of movers' z transition via a normal copula 
with SIPP micro data. This is the implementation described in section C.1 of the online appendix.
}
