% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SippAnalyse.r
\name{SippProbitMove}
\alias{SippProbitMove}
\title{Probit: Determinants of Cross Division Moves}
\usage{
SippProbitMove(d, path = "~/Dropbox/mobility/output/data/sipp")
}
\arguments{
\item{d}{a svy object. Can be obtained by calling \code{\link{SippSvyDesign}}.}

\item{path}{to save output}
}
\value{
produces table 4 in main text: \emph{Determinants of cross census division moves in SIPP data}.
}
\description{
Probit: Determinants of Cross Division Moves
}
\details{
Uses SIPP micro data to estimate a probit model 
relating observables to a indicator of whether a move took place in period t.
}
