% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SippPrepare.r
\name{SippSvyDesign}
\alias{SippSvyDesign}
\title{Create SIPP survey Design object}
\usage{
SippSvyDesign(merged = NULL)
}
\description{
Create SIPP survey Design object
}
\details{
Uses variable \code{HHweight} to build a \code{\link{svydesign}} object.
}
