% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotjulia.R
\name{Var.impulse}
\alias{Var.impulse}
\alias{VAR.impulse}
\title{Illustrate translation of Aggregate to Regional Shocks}
\usage{
VAR.impulse(plotpath = "~/Dropbox/research/mobility/output/data/sipp")
}
\arguments{
\item{plotpath}{location to output figure.}
}
\description{
Illustrate translation of Aggregate to Regional Shocks
}
\details{
This produces figure B.4 in the online appendix titled "10 percent shock to Y"
}
